### 2019-01-04 #1153 @ybayle
From the JOSS suggested list of reviewers, we suggest to the editor, these people:
- @ryancoleman for the geometry keyword
- @ngoldbaum for the 3D keyword
- @r-barnes for the evolutionary biology keywords
