### 2019-01-04 afac1d0
spellcheck joss pdf proof
