+++ paper/paper.md
@@ -31,34 +31,34 @@ Processing schemes to mirror and reorder these points are provided to address fu
 # Background
 
 Living organisms display a high diversity of shapes and sizes that are used to characterize ontogenetic stages and phylogenies [@Jones2015; @Guinot2018; @Musa2018].
-Shape similarities are used as phylogenetic signal ---mostly in extinct species--- and, when combined with ecological data, allow describing trophic habits of species [Randau2016].
+Shape similarities are used as phylogenetic signal ---mostly in extinct species--- and, when combined with ecological data, allow describing trophic habits of species [@Randau2016].
 
 Traditional 2D morphometrics, _ie in situ_ or picture measurements of structural element dimensions, have long been favoured to quantitatively compare extant and extinct shapes of organisms [@Adams2013].
-However, part of these measurements did not reflect the structures geometry and for example depended on their relative size.
+However, part of these measurements did not reflect the geometry of the structures and depended for example on their relative size.
 Landmark-based geometric morphometrics in 2D allows separating a form into size and shape within a set of measures.
 In this method, landmarks define comparable points within a set of objects and capture their geometry.
 A Procrustes superimposition then rescales, rotates and translates these landmarks to virtually superimpose all object geometries [@Zelditch2004; @Adams2013]. 
-Each measured variable however is partly independent from other ones on the same structure as all may have been quantified in different 2D views of the same object.
+Each measured variable, however, is partly independent of other ones on the same structure as all may have been quantified in different 2D views of the same object.
 Then, merging morphometric datasets for global analyses may bring noise in the results.
 
 Landmark-based geometric morphometrics in 3D emerged with recent imaging techniques among others X-ray microtomography [@Adams2013]. 
 3D surfaces are created based on scanned organisms and landmarks can be set on all views of interest of a given specimen and belong to the same dataset, with the same reference space.   
-Few free softwares allow digitizing 3D landmarks and semilandmarks, and they do not offer complete tools for subsequent shape and size analyses.
-Most morphometricians use R packages to perform size and shape analyses but input data must already be in package-specific formats that often highly differ from the output formats generated by landmark acquisition softwares.
-Moreover, the raise of intra- and interspecific asymmetry questions involves post-processing steps of morphometric datasets that are not managed by current R packages.
+Few free software solutions allow digitizing 3D landmarks and semilandmarks, and they do not offer complete tools for subsequent shape and size analyses.
+Most morphometricians use R packages to perform size and shape analyses but input data must already be in package-specific formats that often highly differ from the output formats generated by landmark acquisition software products.
+Moreover, the rise of intra- and interspecific asymmetry questions involves post-processing steps of morphometric datasets that are not managed by current R packages.
 
 # ``Scyland3D`` details
 
-``Scyland3D`` can be installed from the GitHub repository https://github.com/ybayle/Scyland3D or from pip https://pypi.org/project/Scyland3D.
+``Scyland3D`` can be installed from pip https://pypi.org/project/Scyland3D or from the Github repository https://github.com/ybayle/Scyland3D.
 
 We introduce ``Scyland3D`` to convert .pts files, exported by the Landmark software (v3.6), into .csv files.
 ``Scyland3D`` first removes duplicated coordinates because landmarks and semilandmarks might be mixed up in the Landmark software.
 We then allow to mirror the landmark and semilandmark coordinates for a subset of 3D objects and to reorder these points.
-This allows performing 3D geometric morphometric analyses between previously asymmetrical objects (Fig. 1).
+This allows performing 3D geometric morphometric analyses between previously asymmetrical objects (Figure 1).
 
[…]
