+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Adams2013,
-author = {Adams, D. C. and Rohlf, F.J. and Slice, D.E.},
+author = {Adams, D. C. and Rohlf, F. J. and Slice, D. E.},
 journal = {Hystrix},
 number = {1},
 pages = {7},
