### 2019-01-09 #1153 @r-barnes
@ybayle : I haven't reviewed for JOSS before, but I'm generally game to give it a go. From a quick glance at your repository I don't see the evolutionary biology tie-in; however, I do a fair amount of computational science, so this likely wouldn't prevent me from reviewing it.
