### 2019-01-11 #1153 @fberio
>
>
> @ybayle : I haven't reviewed for JOSS before, but I'm generally game to give it a go. From a quick glance at your repository I don't see the evolutionary biology tie-in; however, I do a fair amount of computational science, so this likely wouldn't prevent me from reviewing it.

@r-barnes We designed the functions for an evo-devo project but indeed, this is a global method that can be used for many research topics. We did not focus on Evolution in the description but the code was designed to help with 3D landmarks handling.
