### 2019-01-12 #1153 @Kevin-Mattheus-Moerman
@cleardusk @patrikhuber @anilbas @JuliaWinchester would you be interested in reviewing this work for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/about)? The https://joss.readthedocs.io/en/latest/reviewer_guidelines.html focuses on the software and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01153/joss.01153/10.21105.joss.01153.pdf). The review process (here is an example https://github.com/openjournals/joss-reviews/issues/1148) takes place on GitHub in a dedicated issue that I'll open once we have sufficient reviewers. Let me know if you have questions.
