### 2019-01-12 #1153 @patrikhuber
@Kevin-Mattheus-Moerman I can't currently do a full review according to your reviewer guidelines, but I've looked at the paper & repo and I'd be happy to leave a paragraph or two of review comments.
