### 2019-01-31 #1153 @Kevin-Mattheus-Moerman
@cleardusk @anilbas @JuliaWinchester would you be interested in reviewing this work for the Journal of Open Source Software (JOSS)? The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses on the software and a short paper. The review process (here is an example #1148) takes place on GitHub in a dedicated issue that I'll open once we have sufficient reviewers. Let me know if you have questions.
