### 2019-02-10 #1153 @patrikhuber
@Kevin-Mattheus-Moerman Apologies for a bit of a delay. Here goes my review:

### Overview

It seems to me that in general, this software is for a very, _**very**_ limited use case (which of course isn't necessarily bad!). Basically, the main thing it does:
1) Reads `.pts` landmarks from a software called "Landmark software (v3.6)" and converts them to a `.csv` format
2) It can mirror and re-order those landmarks while converting them. The mirroring potentially seems like it can be a bit tricky (see more below).

### Comments/questions:

* The Readme needs a better introduction in my opinion. It states

> Scyland3D is a Python tool for converting 3D raw landmark and semilandmark coordinates to a usable format for geometric morphometric analyses

but if you're not from _exactly_ that field, you're going to have a hard time discerning if the tool is useful or not, because "3D raw landmarks" is such a large field. I am not from the author's field directly but "geometric morphometric analysis" looks like it is in fact what I know by the term shape analysis, which is a really broad field, with thousands of different landmark formats. And Sciland3D seems only for one specific task in one very particular field (or perhaps even one particular lab or commercial software), without much broader applicability, I think. At least for me as a computer vision researcher, having worked with statistical shape models a lot, I had a hard time figuring out what this exactly is and does.

* The PDF and repo don't explain or introduce what "semilandmarks" are. I've worked with facial landmark points for many years and don't believe I've come across this term. Also no example seems to make use of them, as far as I can tell. Some explanation here would be really great.

* The Readme mentions `.pts` files but not what they are or where they're coming from. We have `.pts` files too in the facial landmarks community and they are something completely different (e.g. see https://github.com/patrikhuber/eos/blob/master/examples/data/image_0010.pts).
The PDF says they're ".pts files, exported by the Landmark software (v3.6)" but it also doesn't help much - please add a link/reference to that "Landmark software (v3.6)" or even better, to the format definition as well.

To sum that part up, I think there's way more context needed, and the name "Scyland3D: Processing 3D landmarks" may also be too general.

### Specific comments regarding the paper (PDF):

* The abstract says "convert them to a usable format". Why would the pts files be a less usable format than csv files? I believe that might be the wrong motivation or the motivation is just misstated. Both formats are probably useful, but data conversion is part of our daily jobs so we just want to to be as easy, painless and error-free as possible.

* I believe "ie" should have periods, i.e. (no pun intended... :-) ) "i.e.". Also I wouldn't italicise it but I'm not sure whether that's perhaps just personal preference. Italicising is for emphasising and what you want to emphasise there is _in situ_ and not the "i.e." :-) (see e.g. https://english.stackexchange.com/questions/3911/correct-spelling-italicization-of-e-g-i-e)

* _"Few free software solutions allow digitizing 3D landmarks and semilandmarks [...]"_: What do you mean with that? Store them to the disk? I think again what you probably mean is few ready-made scripts exist that convert between all these many formats?

* _"checks and reports [...] Consistency of folder and file names"_ In what way does it check that, what naming scheme does it expect?

* _"checks and reports [...] Precision of landmark and semilandmark coordinates export by the Landmark software."_ - I can't find anything about that in the code?

* _"mirrored objects have reverted bending"_ What do you mean with blending here?

### Specifically with regards to the code:

* How useful is the mirroring? For it to be useful, you usually have to mirror the original data too (in your case e.g. the teeth point-cloud), in the same way. Furthermore it seems to me that e.g. in the example given in Figure 1, this only works if the landmarks are placed in a symmetric way in the first place (i.e. if 1 is the "mirror-point" of 7, etc.).

* The code could use more documentation in general. A tiny bit longer function documentation, but also in-code comments. Code should be either self-explanatory and need no comments (best case) or have appropriate comments, where needed.

* Here it says something about a comparison "up to epsilon": https://github.com/ybayle/Scyland3D/blob/1d219d40a8327515d936b94a2d1d4ce7060c0f16/Scyland3D.py#L80
But in the following code, I can't find any comparison up to an epsilon, only equality checks. It may very well be that I don't read the code correctly and I haven't run it either but it kind of reinforces my point from above about documentation.

* https://github.com/ybayle/Scyland3D/blob/1d219d40a8327515d936b94a2d1d4ce7060c0f16/Scyland3D.py#L106-L107
It would be nice to know here what's "good" about one link and "wrong" about one link. There seem to be some peculiarities at play here, and even a StackOverflow answer with most upvotes and comments like "This one should be marked as the correct answer." that apparently has gotten something important wrong. Let us know about it please, and also again, please put comments in that function and let us know which ones were/are the tricky bits.
From what I can see, that function (`reverse_z`) is actually one of *the* main contributions of the proposed submission.

Some of those comments might be a bit picky but in particular because I don't think this contribution is particularly big, it is even more important that the small thing that *is* contributed is as useful as possible and can be used by people being assured it doesn't have any bugs, or if they do need to check something, the code should be documented clearly enough to make it easy.
