### 2019-02-18 #1153 @r-barnes
Sorry about the delay. Whedon invited me to join the github and this resulted in an unbelievable amount of spam. I think I've finally got my settings such that I'll see this conversation and not the rest. @Kevin-Mattheus-Moerman : Do I now evaluate the work based on [this](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and [this](https://joss.readthedocs.io/en/latest/review_criteria.html)? (In my mind there was some kind of standard form with checkboxes and spaces for comments.)
