### 2019-02-18 #1262 @Kevin-Mattheus-Moerman
@ybayle this is where the review takes place. Please also note the comments by @patrikhuber over at https://github.com/openjournals/joss-reviews/issues/1153
