### 2019-03-07 #1262 @labarba
I notice that the `.py` file mentions MIT License, while the `LICENSE` file is GNU.
https://github.com/ybayle/Scyland3D/blob/1d219d40a8327515d936b94a2d1d4ce7060c0f16/Scyland3D.py#L5
