### 2019-03-20 #1262 @ybayle
> I notice that the `.py` file mentions MIT License, while the `LICENSE` file is GNU.
> https://github.com/ybayle/Scyland3D/blob/1d219d40a8327515d936b94a2d1d4ce7060c0f16/Scyland3D.py#L5

Thanks for noticing! We have updated our license consistently accordingly.
