### 2019-03-27 ybayle/Scyland3D#1 @r-barnes
Running without arguments fails
Configuration shows a command-line entry point
```
    entry_points='''
        [console_scripts]
        Scyland3D=Scyland3D:pts2csv
    ''',
```

Running `Scyland3D` on the command line raises an error:
```
rick@sally:/z$ Scyland3D
Traceback (most recent call last):
  File "/home/rick/.local/bin/Scyland3D", line 11, in <module>
    sys.exit(pts2csv())
TypeError: pts2csv() missing 1 required positional argument: 'indir'
```
