### 2019-03-27 ybayle/Scyland3D#1 @r-barnes
It seems us that the program is adapted only for use within the Python environment, despite having this command line functionality. I suggest that you flesh out the command line functionality using the argparse package. This is likely to expand the scope of users that can benefit from your work.
