### 2019-03-27 ybayle/Scyland3D#8 @r-barnes
Comments in code
The source code tells the computer what to do, but does not explain what you intended the computer to do. Given the speed with which the open-source world changes, and the lack of incentives to maintain software with academia, I think it's valuable to have comments within code so that people providing patches, new features, or updating the code have an easy way of orientating themselves to the intended functionality of what can otherwise be mysterious code blocks.

Please add additional comments within the code.
