### 2019-03-27 #1262 @r-barnes
I've completed the JOSS rubric and have raised concerns linked above.
