### 2019-03-27 #1262 @r-barnes
(I want to apologize for the delay in getting to this review, I've been experiencing some issues with RSI which is made it difficult to keep up with computer-based tasks. I'm now using a speech recognition program to do the review, so I apologize for any grammatical  errors or odd choices of words.)

Many of the issues that I've raised are on the target repository.

I share @patrikhuber's concerns that this piece of software is super specific. When I saw the name of the software and description of the softer, I expected to find a large amount of code relating to the manipulation of 3D points, instead the softer it appears to consist predominantly of what's essentially a utility function for a much larger pipeline. Given the limited scope of the software, I would have difficulty recommending it for publication.

Similar to the other reviewer, I don't see any kind of epsilon cutoff to deal with floating-point issues. I also worry about the utility of the mirroring, it seems as though the user has to specify the plane of the mirror. it seems to me that this is a problem that is better suited to a technique such as iterative closest point or optimal transport.

The code is at times verbose:
```
coord = []
    for item in data:
        tmp = []
        for xyz in item.split(","):
            tmp.append(float(xyz))
coord.append(tmp)
```
could be rendered as
```
coord = [float(c) for item in data for c in item.split(",")]
```
Granted, such techniques should not be employed where they impede the readability of code, but I think there are several instances in which the existing code is more complex than is necessary. `export2csv()` might be the best example of this, since Python includes an entire core module for solving exactly this problem.

This contribution is small enough that it should be as perfect and rock-solid as possible. For that, I'm looking for cleaner, better-commented code, test cases, and an exceptionally well-explained use case. Even so, I think this function is just too limited in scope to make a good contribution.
