### 2019-03-30 614e9a5
add python 2 and 3 compatibility message #4
