### 2019-03-30 e83a1c8
enhance community guidelines #2
