+++ CONTRIBUTING.md
@@ -1,5 +1,5 @@
 # Contributing
 
 Contributions are welcome!
-If you found a bug, please submit an issue [here](https://github.com/ybayle/Scyland3D/issues/new).
-If you are willing to propose new features, please submit a pull request [here](https://github.com/ybayle/Scyland3D/pulls).
+If you found a bug or need support, please submit an issue [here](https://github.com/ybayle/Scyland3D/issues/new).
+If you are willing to propose new features or have bug fixes to contribute, please submit a pull request [here](https://github.com/ybayle/Scyland3D/pulls).
