+++ CHANGELOG.md
@@ -1,0 +1,16 @@
+# CHANGELOG.md
+
+## 1.0.2 2018-03-30
+
+add change log
+add arguments parsing #1
+
+## 1.0.1 2018-03-30
+
+enhance community guidelines #2
+add python 2 and 3 compatibility message #4
+use python core csv module #9 
+
+## 1.0.0 2018-01-03
+
+First release
