### 2019-03-31 ybayle/Scyland3D#1 @ybayle
Thanks for this suggestion. The arg parsing is now possible.
