### 2019-04-10 #1262 @ybayle
Yes, I have started to fix some issues reported by the reviewers on the main repo! I'll keep you updated.
