### 2019-06-12 ybayle/Scyland3D#8 @ybayle
Additional comments and description were added to the code. Tell me if some part of code deserve more description, thanks.
