### 2019-06-12 8827c5e
fix #3 #7 #8 #9
