+++ CHANGELOG.md
@@ -1,16 +1,20 @@
 # CHANGELOG.md
 
-## 1.0.2 2018-03-30
+## 1.0.3 2019-06-12
+
+Python refactor and added comment in the code #3 #7 #8 #9
+
+## 1.0.2 2019-03-30
 
 Add change log
 Add arguments parsing #1
 
-## 1.0.1 2018-03-30
+## 1.0.1 2019-03-30
 
 Enhance community guidelines #2
 Add python 2 and 3 compatibility message #4
 Use python core csv module #9 
 
-## 1.0.0 2018-01-03
+## 1.0.0 2019-01-03
 
 First release
