### 2019-06-12 53955e4
add image for description of the process #5
