### 2019-06-12 #1262 @ybayle
@labarba: Sorry for the delay, I have taken into account most comments from @r-barnes in the last commits and started processing comments from @patrikhuber locally that will be pushed soon.
