### 2019-06-13 168dd2a
Add valid python docstring #6
