### 2019-06-13 #1262 @ybayle
@patrikhuber
Thank you for your valuable comments to improve our submission, we detail below our changes.

> The Readme needs a better introduction in my opinion.
> The PDF and repo don't explain or introduce what "semilandmarks" are.

The ReadMe and the paper have been updated accordingly.

> The Readme mentions .pts files but not what they are or where they're coming from. We have .pts files too in the facial landmarks community and they are something completely different (e.g. see https://github.com/patrikhuber/eos/blob/master/examples/data/image_0010.pts).

As I can see, we provide as examples nearly the same .pts files except, we have an additional column for values for the z-axis. We added a link in the paper and in the ReadMe to the definition of the .pts file format we use.

> "Few free software solutions allow digitizing 3D landmarks and semilandmarks [...]": What do you mean with that? Store them to the disk? I think again what you probably mean is that few ready-made scripts exist that convert between all the many formats?

You raise an interesting question there about the variety of vocabulary used here. We saw in numerous places (e.g. https://www.rdocumentation.org/packages/geomorph/versions/3.1.2/topics/digitize2d or http://people.tamu.edu/~alawing/materials/ESSM689/Quick_Guide_to_Geomorph_v2.0.pdf) the word "digitizing" used for "data acquisition". In order to avoid confusion with other readers we used both words in the paper.

> "mirrored objects have reverted bending" What do you mean with bending here?
> How useful is the mirroring? For it to be useful, you usually have to mirror the original data too (in your case e.g. the teeth point-cloud), in the same way. Furthermore it seems to me that e.g. in the example given in Figure 1, this only works if the landmarks are placed in a symmetric way in the first place (i.e. if 1 is the "mirror-point" of 7, etc.).

We are talking here about the main curvature a surface can display in terms of how much they display a concave or convex shape on a given axis. In our example with shark's teeth, they are longer (x) than wide (y) and very thin (z). Teeth are concave on the z-axis and their tip are not centred relatively to their base. So, when you mirror them to match left teeth to right teeth, you have their base and tip that are overlapping but you one is concave while the other is convex as can be seen on their z-axis. That's why "mirrored objects have reverted bending".

> The code could use more documentation in general. A tiny bit longer function documentation, but also in-code comments. Code should be either self-explanatory and need no comments (best case) or have appropriate comments, where needed.

We had the same comment from the second reviewer and updated our code accordingly.

> Here it says something about a comparison "up to epsilon": https://github.com/ybayle/Scyland3D/blob/1d219d40a8327515d936b94a2d1d4ce7060c0f16/Scyland3D.py#L80 But in the following code, I can't find any comparison up to an epsilon, only equality checks. It may very well be that I don't read the code correctly and I haven't run it either but it kind of reinforces my point from above about documentation.

Indeed, the first version of the script was using an epsilon as a workaround to report duplicated landmarks in the exported .pts files. A better solution has been found since. The comment is now updated.
