### 2019-06-13 ec26966
fix patrikhuber's comments about increasing details in the code and paper
