+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Abstract
 
-``Scyland3D`` is a Python tool for converting 3D raw landmark and semilandmark coordinates to a usable format for geometric morphometric analyses.
+``Scyland3D`` is a Python tool for converting 3D raw landmark and semilandmark coordinates exported from landmark acquisition softwares to a csv format best suited for geometric morphometric analyses.
 Processing schemes to mirror and reorder these points are provided to address further symmetry issues.
 This is useful when working on left-right differences in a given species. 
 This tool is designed to pre-process the raw extracted landmarks and to generate a database-like CSV file for statistical analyses.
@@ -36,7 +36,7 @@ The main targeted audience are biologist looking for an easy and quick parser fo
 Living organisms display a high diversity of shapes and sizes that are used to characterize ontogenetic stages and phylogenies [@Jones2015; @Guinot2018; @Musa2018].
 Shape similarities are used as phylogenetic signal ---mostly in extinct species--- and, when combined with ecological data, allow describing trophic habits of species [@Randau2016].
 
-Traditional 2D morphometrics, _ie in situ_ or picture measurements of structural element dimensions, have long been favoured to quantitatively compare extant and extinct shapes of organisms [@Adams2013].
+Traditional 2D morphometrics, _i.e. in situ_ or picture measurements of structural element dimensions, have long been favoured to quantitatively compare extant and extinct shapes of organisms [@Adams2013].
 However, part of these measurements did not reflect the geometry of the structures and depended for example on their relative size.
 Landmark-based geometric morphometrics in 2D allows separating a form into size and shape within a set of measures.
 In this method, landmarks define comparable points within a set of objects and capture their geometry.
@@ -46,7 +46,8 @@ Then, merging morphometric datasets for global analyses may bring noise in the r
 
 Landmark-based geometric morphometrics in 3D emerged with recent imaging techniques among others X-ray microtomography [@Adams2013]. 
 3D surfaces are created based on scanned organisms and landmarks can be set on all views of interest of a given specimen and belong to the same dataset, with the same reference space.   
-Few free software solutions allow digitizing 3D landmarks and semilandmarks, and they do not offer complete tools for subsequent shape and size analyses.
+Among softwares allowing digitizing (data acquisition of) 3D landmarks and semilandmarks, there are few free softwares and they do not offer complete tools for subsequent shape and size analyses.
+A landmark is a point set up by hand as opposed to a semilandmark that is interpolated by the computer between two landmarks by following the curvature of the studied form.
 Most morphometricians use R packages to perform size and shape analyses but input data must already be in package-specific formats that often highly differ from the output formats generated by landmark acquisition software products.
 Moreover, the rise of intra- and interspecific asymmetry questions involves post-processing steps of morphometric datasets that are not managed by current R packages.
 
@@ -54,7 +55,7 @@ Moreover, the rise of intra- and interspecific asymmetry questions involves post
 
 ``Scyland3D`` can be installed from pip https://pypi.org/project/Scyland3D or from the Github repository https://github.com/ybayle/Scyland3D.
 
-We introduce ``Scyland3D`` to convert .pts files, exported by the Landmark software (v3.6), into .csv files.
+We introduce ``Scyland3D`` to convert .pts files (format define on page 37 in http://www.idav.ucdavis.edu/research/projects/EvoMorph/supplement/LandmarkDoc_v3_b6.pdf), exported by the Landmark software (v3.6 cf http://www.idav.ucdavis.edu/research/projects/EvoMorph/supplement/LandmarkDoc_v3_b6.pdf) or any landmark processing software, into .csv files.
[…]
