### 2019-06-13 #1262 @ybayle
@r-barnes

> I also worry about the utility of the mirroring, it seems as though the user has to specify the plane of the mirror. it seems to me that this is a problem that is better suited to a technique such as iterative closest point or optimal transport.

We detailed further our code to explain why the mirroring is needed. We also added details about the fact that our code finds a plane and that no user specified plane has to be supplied.
