+++ Scyland3D.py
@@ -71,10 +71,7 @@ def _remove_duplicates(data):
     # Remove duplicates because some landmarks and semilandmarks are sometimes exported twice.
     coord = []
     for item in data:
-        tmp = []
-        for xyz in item.split(","):
-            tmp.append(float(xyz))
-        coord.append(tmp)
+        coord.append([float(xyz) for xyz in item.split(",")])
     index_to_remove = []
     for k in range(3):
         for i in range(len(coord) - 1):
@@ -95,7 +92,7 @@ def _remove_duplicates(data):
 def _reverse_z(data):
     """reverse_z
     Find the relative z-axis for a given set of point by computing a plane that minimizes the difference between it and the given points.
-    Then reverts the set of points relatively to this new plane.
+    Then reverts the set of points relatively to this new plane, so no user specific plane has to be supplied.
     Inspired from https://www.youtube.com/watch?v=86lwcXeZoiA
     The suggestion from https://stackoverflow.com/questions/8954326/how-to-calculate-the-mirror-point-along-a-line#8954454 is not applicable here as it is not valid in 3D and does not take into account relative axes. 
 
