+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name = 'Scyland3D',
-    version = '1.0.2',
+    version = '1.0.3',
     description = 'A Python package for processing 3D landmarks',
     long_description=open('README.md').read(),
     author = 'Fidji Berio and Yann Bayle',
