### 2019-06-14 #1262 @patrikhuber
Hi @ybayle,

Thanks for addressing the comments! I think it looks much better!

A few things that I think still warrant clarification, or that I've noticed while reviewing the changes:

* > Readme: "...exported from landmark acquisition softwares..."

It still looks to me like your script is very specific, more concretely, it seems specific to the "Landmark 3.6", "Landmark Editor" software from IDAV, UC Davis (http://www.idav.ucdavis.edu/research/EvoMorph).
Sure, other programs may export the exact same .pts files (do you know of any such?), but I think it would clarify the context of your script if this was prominently mentioned in the Readme introduction, and it would help people using the "Landmark 3.6" software to find your script and know that it might be useful to them.

* [In this image](https://github.com/ybayle/Scyland3D/raw/master/paper/figure1.png), don't you want to mirror the landmark numbers as well? Otherwise you change the semantic meaning of each landmark? In my field, when mirroring, we want to preserve the semantic meaning of a landmark when mirroring and not reorder them but it might be different in your use field.

* > "Scyland3D is licensed under the MIT License as described in the license file. Scyland3D was mainly created for research purposes and thus can be used freely for research and academic use with the following citation: [...]"

I'm not sure if the wording of the second sentence could be a bit confusing. It may be interpreted as a restriction on the MIT License? IANAL but perhaps rewording might make sense.

* https://github.com/ybayle/Scyland3D/blob/ba701ad3e73e5a1b0dbf69f8dfb1be4b3dcf8710/Scyland3D.py#L95
"reverts": Not sure I understand. Do you perhaps mean "reverses"?

* https://github.com/ybayle/Scyland3D/blob/ba701ad3e73e5a1b0dbf69f8dfb1be4b3dcf8710/Scyland3D.py#L80
Do you check here if the x, y and z coordinates are the same, or only one of them?

* > "Scyland3D first removes duplicated coordinates because landmarks and
semilandmarks might be mixed up in the Landmark software."

"might be mixed up" is not very clear/scientific. Is this a particular problem/bug in the "Landmark 3.6" software?
It seems to me that if it mixes up landmarks and semilandmarks, that just removing duplicates would not really be a solution?

* Throughout the repo/paper: I think "software" is an uncountable noun and you can't (or shouldn't) use "softwares". In most cases you can just use the singular (or "applications", "pieces of software" etc.). Also there were a couple of instances of dodgy grammar/wording - I'd advise proofreading by a native speaker.

* I think you might be mixing BE/AE? E.g. "characterizing" vs "favoured".

* I found a few more typos across the Readme and Paper (didn't write them down sorry - one of them was "biologist"=>"biologists" in the paper).
