### 2019-06-23 15480e3
fix comment from patrikhuber issue 1262 from joss
