+++ paper/paper.md
@@ -25,15 +25,15 @@ bibliography: paper.bib
 
 # Abstract
 
-``Scyland3D`` is a Python tool for converting 3D raw landmark and semilandmark coordinates exported from landmark acquisition softwares to a csv format best suited for geometric morphometric analyses.
+``Scyland3D`` is a Python tool for converting 3D raw landmark and semilandmark coordinates exported from landmark acquisition software to a csv format best suited for geometric morphometric analyses.
 Processing schemes to mirror and reorder these points are provided to address further symmetry issues.
 This is useful when working on left-right differences in a given species. 
 This tool is designed to pre-process the raw extracted landmarks and to generate a database-like CSV file for statistical analyses.
-The main targeted audience are biologist looking for an easy and quick parser for raw landmarks.
+The main targeted audience are biologists looking for an easy and quick parser for raw landmarks.
 
 # Background
 
-Living organisms display a high diversity of shapes and sizes that are used to characterize ontogenetic stages and phylogenies [@Jones2015; @Guinot2018; @Musa2018].
+Living organisms display a high diversity of shapes and sizes that are used to characterise ontogenetic stages and phylogenies [@Jones2015; @Guinot2018; @Musa2018].
 Shape similarities are used as phylogenetic signal ---mostly in extinct species--- and, when combined with ecological data, allow describing trophic habits of species [@Randau2016].
 
 Traditional 2D morphometrics, _i.e. in situ_ or picture measurements of structural element dimensions, have long been favoured to quantitatively compare extant and extinct shapes of organisms [@Adams2013].
@@ -42,11 +42,11 @@ Landmark-based geometric morphometrics in 2D allows separating a form into size 
 In this method, landmarks define comparable points within a set of objects and capture their geometry.
 A Procrustes superimposition then rescales, rotates and translates these landmarks to virtually superimpose all object geometries [@Zelditch2004; @Adams2013]. 
 Each measured variable, however, is partly independent of other ones on the same structure as all may have been quantified in different 2D views of the same object.
-Then, merging morphometric datasets for global analyses may bring noise in the results.
+Then, merging morphometric datasets for global analyses may bring noises in the results.
 
-Landmark-based geometric morphometrics in 3D emerged with recent imaging techniques among others X-ray microtomography [@Adams2013]. 
+Landmark-based geometric morphometrics in 3D emerged with recent imaging techniques among other X-ray microtomography [@Adams2013]. 
 3D surfaces are created based on scanned organisms and landmarks can be set on all views of interest of a given specimen and belong to the same dataset, with the same reference space.   
-Among softwares allowing digitizing (data acquisition of) 3D landmarks and semilandmarks, there are few free softwares and they do not offer complete tools for subsequent shape and size analyses.
+Among software allowing digitizing (data acquisition of) 3D landmarks and semilandmarks, there are few free software and they do not offer complete tools for subsequent shape and size analyses.
 A landmark is a point set up by hand as opposed to a semilandmark that is interpolated by the computer between two landmarks by following the curvature of the studied form.
 Most morphometricians use R packages to perform size and shape analyses but input data must already be in package-specific formats that often highly differ from the output formats generated by landmark acquisition software products.
 Moreover, the rise of intra- and interspecific asymmetry questions involves post-processing steps of morphometric datasets that are not managed by current R packages.
@@ -55,8 +55,8 @@ Moreover, the rise of intra- and interspecific asymmetry questions involves post
 
 ``Scyland3D`` can be installed from pip https://pypi.org/project/Scyland3D or from the Github repository https://github.com/ybayle/Scyland3D.
 
[…]
