### 2019-06-23 #1262 @ybayle
Hi @patrikhuber,

Thanks for your comments.

> It still looks to me like your script is very specific, more concretely, it seems specific to the "Landmark 3.6", "Landmark Editor" software from IDAV, UC Davis (http://www.idav.ucdavis.edu/research/EvoMorph).

I have updated the ReadMe to prominently mention this in the introduction.

> In this image, don't you want to mirror the landmark numbers as well? Otherwise you change the semantic meaning of each landmark? In my field, when mirroring, we want to preserve the semantic meaning of a landmark when mirroring and not reorder them but it might be different in your use field.

Indeed, our script handle the inversion or not of the numbering of the landmarks. Our use-case is closely related to this kind of processing, thus justifying our current example image.

> It may be interpreted as a restriction on the MIT License?

I have reworded the license.

> "reverts": Not sure I understand. Do you perhaps mean "reverses"?

Thanks you for pointing out this typo.

> Do you check here if the x, y and z coordinates are the same, or only one of them?

This sentence has been reworded.

> "might be mixed up" is not very clear/scientific.

This has been reworded. Indeed, the landmark acquisition software avoid setting landmarks at the same position but landmarks can be set up at the same position of a semilandmark, which is not wanted.

> Also there were a couple of instances of dodgy grammar/wording. I think you might be mixing BE/AE? E.g. "characterizing" vs "favoured".

The paper and ReadMe have been checked for misspelling and now use BE. Please confirm no mistakes or typos are remaining.
