### 2019-06-23 #1262 @patrikhuber
Hi @ybayle,

Looks good!

> Indeed, our script handle the inversion or not of the numbering of the landmarks.

I see - but I'm not sure how? Could this be the `order_factor` ? The example in the Readme sets this to `'upper'`. But to be honest I really can't figure out what it might be doing - the code doesn't seem to check for any particular value like 'upper'? What are the valid values, and what do they each do? This could be documented better (e.g. also added to the "help" option).

> `# Reorder the landmarks to avoir bias during landmarks set up by humans`

Typo "avoir" :) Also the meaning is not clear to me - "...during landmarks set up by humans"?

Last thing, `feature_names` is listed in the Readme as an argument to `pts2csv` but a CTRL+F of the code yields no results and it's also not part of the function signature?
