### 2019-06-24 #1262 @ybayle
Hi @patrikhuber,

> Looks good!

Thank you again for your comments and suggestions that help improve a lot our submission and our code. We also learn many python-related things during this process!

> I see - but I'm not sure how? Could this be the order_factor ? The example in the Readme sets this to 'upper'. But to be honest I really can't figure out what it might be doing - the code doesn't seem to check for any particular value like 'upper'? What are the valid values, and what do they each do? This could be documented better (e.g. also added to the "help" option).

If the dataset contains lower and upper teeth where only the upper teeth need to be reordered, one can use: `Scyland3D.pts2csv("example/", order_factor="upper", order=[36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 37])` or from the command line: `python Scyland3D.py -i "example/" -f "upper" -o "36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 37"`. This is only valid if the keyword `upper` is present in the file name. Information regarding this feature have been added to the ReadMe and to the code.

> Last thing, feature_names is listed in the Readme as an argument to pts2csv but a CTRL+F of the code yields no results and it's also not part of the function signature?

This is fixed and we also added better example in the ReadMe to describe how to use our package from another python script or from the command line.
