### 2019-06-24 3626b2a
refactor for command line support and add more comments in the code and ReadMe
