+++ Scyland3D.py
@@ -4,8 +4,8 @@
 # E-mails   fidji.berio@ens-lyon.fr and bayle.yann@live.fr
 # License   MIT
 # Created   15/02/2018
-# Updated   13/06/2019
-# Version   1.0.3
+# Updated   24/06/2019
+# Version   1.0.4
 #
 
 import os
@@ -15,7 +15,7 @@ import getopt
 import numpy as np
 
 
-def _export2csv(data, nb_landmark, indir, modif=""):
+def _export2csv(data, nb_landmark, indir, feature_names=None, modif=""):
     """export2csv
     Export data to a CSV named indir + "../landmarks" + modif + ".csv".
 
@@ -23,16 +23,25 @@ def _export2csv(data, nb_landmark, indir, modif=""):
         data (array): The landmarks to export.
         nb_landmark (int): The number of landmarks.
         indir (str): The name of the directory where the output files will be stored.
-        modif (str): The name of the modification applied to the data.
+        feature_names (str or array of str): String describing each feature names separated with a comma or array of string (e.g. "age,sex,size" if supplied from the command line or ["age", "sex", "size"] if supplied in python script).
+        modif (str): The name of the modification applied to the data (e.g. none, reversed, and/or reordered).
     """
     # Generate the header of the csv
     fieldnames = ["ID"]
+    assert nb_landmark > 0, "The number of landmarks must be positive."
     for numb in range(1, nb_landmark + 1):
         for axe in ["x", "y", "z"]:
             fieldnames.append(axe + str(numb))
     nb_feature = len(data[0]) - nb_landmark * 3 - 1
-    for numb in range(1, nb_feature + 1):
-        fieldnames.append("Feature" + str(numb))
+    if feature_names is not None:
+        if isinstance(feature_names, str):
+            feature_names = feature_names.split(",")
+        assert nb_feature == len(feature_names), "The number of feature names provided (" + str(len(feature_names)) + ") does not match the number of features detected in the files " + str(nb_feature) + "."
+        for feature_name in feature_names:
+            fieldnames.append(feature_name)
+    else:
+        for numb in range(1, nb_feature + 1):
+            fieldnames.append("Feature" + str(numb))
     # Export the header and the data  
     with open(indir + "../landmarks" + modif + ".csv", "w") as out_file:
         csv.DictWriter(out_file, fieldnames=fieldnames).writeheader()
@@ -131,19 +140,19 @@ def _reverse_z(data):
     return data_mirror
 
 
-def pts2csv(indir="example/", mirror_factor=None, order=None, order_factor=None, verbose=True):
+def pts2csv(indir="example/", mirror_factor=None, order=None, order_factor=None, feature_names=None, verbose=True):
     """pts2csv
     Convert .pts files from indir to a single .csv file
 
     Args:
         indir (str): The name of the directory where the output files will be stored.
         mirror_factor (str): The name of the factor to use for mirroring the landmarks on the z-axis.
-        order (array): An array of int indicating how to reorder the landmarks.
+        order (str or array of int): A string (if supplied by the command line) or an array of int (if called from another python script) indicating how to reorder the landmarks (e.g. "1,3,2" if supplied from the command line or [1, 3, 2] if supplied by another python script).
         order_factor (str): The name of the factor to use for reordering the landmarks.
         verbose (bool): Whether to output details during the process.
     """
     assert os.path.exists(indir) and os.path.isdir(indir), indir + " not found."
-    assert (order is None and order_factor is None) or (order is not None and order_factor is not None), "Must supply order and order_factor."
[…]
