### 2019-06-24 #1262 @patrikhuber
You're welcome, glad it is helpful.

I understand, thanks for the additional explanations.

I think you forgot to document the `feature_names` here in the "Args" list:
https://github.com/ybayle/Scyland3D/blob/3626b2ad4c339ee2f2483d5a0c06dd94df96f615/Scyland3D.py#L151-L152

Also you should format the .py file with PEP 8, there are a few lines that are too long (and potentially some other things). Any Python IDE should be able to do it, in PyCharm it's "Code => Reformat Code".

Otherwise it looks all good from my side! 👍
