+++ Scyland3D.py
@@ -4,8 +4,8 @@
 # E-mails   fidji.berio@ens-lyon.fr and bayle.yann@live.fr
 # License   MIT
 # Created   15/02/2018
-# Updated   24/06/2019
-# Version   1.0.4
+# Updated   25/06/2019
+# Version   1.0.5
 #
 
 import os
@@ -23,8 +23,11 @@ def _export2csv(data, nb_landmark, indir, feature_names=None, modif=""):
         data (array): The landmarks to export.
         nb_landmark (int): The number of landmarks.
         indir (str): The name of the directory where the output files will be stored.
-        feature_names (str or array of str): String describing each feature names separated with a comma or array of string (e.g. "age,sex,size" if supplied from the command line or ["age", "sex", "size"] if supplied in python script).
-        modif (str): The name of the modification applied to the data (e.g. none, reversed, and/or reordered).
+        feature_names (str or array of str): String describing each feature names separated with a
+            comma or array of string (e.g. "age,sex,size" if supplied from the command line or
+            ["age", "sex", "size"] if supplied in python script).
+        modif (str): The name of the modification applied to the data (e.g. none, reversed, and/or
+            reordered).
     """
     # Generate the header of the csv
     fieldnames = ["ID"]
@@ -36,13 +39,19 @@ def _export2csv(data, nb_landmark, indir, feature_names=None, modif=""):
     if feature_names is not None:
         if isinstance(feature_names, str):
             feature_names = feature_names.split(",")
-        assert nb_feature == len(feature_names), "The number of feature names provided (" + str(len(feature_names)) + ") does not match the number of features detected in the files " + str(nb_feature) + "."
+        assert nb_feature == len(feature_names), (
+            "The number of feature names provided ("
+            + str(len(feature_names))
+            + ") does not match the number of features detected in the files "
+            + str(nb_feature)
+            + "."
+        )
         for feature_name in feature_names:
             fieldnames.append(feature_name)
     else:
         for numb in range(1, nb_feature + 1):
             fieldnames.append("Feature" + str(numb))
-    # Export the header and the data  
+    # Export the header and the data
     with open(indir + "../landmarks" + modif + ".csv", "w") as out_file:
         csv.DictWriter(out_file, fieldnames=fieldnames).writeheader()
         csv.writer(out_file).writerows(data)
@@ -59,7 +68,9 @@ def _list_pts(indir):
         An array of .pts file names to process. 
     """
     assert os.path.exists(indir) and os.path.isdir(indir), indir + " not found."
-    filenames = [os.path.join(dp, f) for dp, dn, fn in os.walk(os.path.expanduser(indir)) for f in fn]
+    filenames = [
+        os.path.join(dp, f) for dp, dn, fn in os.walk(os.path.expanduser(indir)) for f in fn
+    ]
     filenames = [filen for filen in filenames if ".pts" in filen]
     assert filenames, "There are no .pts files in " + indir
     return filenames
@@ -86,10 +97,13 @@ def _remove_duplicates(data):
         for i in range(len(coord) - 1):
             for j in range(len(coord) - i - 1):
                 if coord[i][k] == coord[i + j + 1][k]:
-                    # If two landmarks share at least two similar coordinates, then the rows are considered as duplicates and the second one will be removed.
+                    # If two landmarks share at least two similar coordinates, then the rows are
+                    # considered as duplicates and the second one will be removed.
                     tmp = [0, 1, 2]
                     tmp.remove(k)
[…]
