### 2019-06-25 #1262 @ybayle
Hi @patrikhuber,
Everything should be fine now and I have formatted the code using https://github.com/python/black.
Thanks again for your comments!
