### 2019-06-25 4f7bbcb
update version number in config file
