### 2019-06-25 9426a2e
update setup for pypi upload v1.0.5
