+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include test/*
+include example/*/*
