### 2019-06-25 #1262 @ybayle
Hi @r-barnes: the new version v1.0.6 with all the fixes is now available on pypi https://pypi.org/project/Scyland3D/
