### 2019-06-26 #1262 @patrikhuber
@ybayle Just fyi the tooth image doesn't load on the pypi page, at least in my browser. If you have the problem as well, it might be because of the relative URL. I just checked my project and what I did was use the "raw" URL for the image (I probably found that solution in stackoverflow or somewhere): E.g.: `https://raw.githubusercontent.com/patrikhuber/eos/gh-pages/images/sfm_shape_3448_mesh.png`
