### 2019-06-26 682b796
fix img abs path for pypi compatibility
