### 2019-06-26 #1262 @ybayle
Hi @patrikhuber,
The pypi repo now displays correctly the image thanks to your suggested url!
