### 2019-06-26 #1262 @ybayle
@Kevin-Mattheus-Moerman Could you please update the version to https://github.com/ybayle/Scyland3D/releases/tag/v1.0.16 in the description?
