### 2019-08-01 #1262 @patrikhuber
@Kevin-Mattheus-Moerman I don't mind personally. But if there's anything more that would help you (like doing an official checklist), I'd be happy to do it. Should be trivial given I've already completed a pretty much full review and everything is fine from my end.
I am on vacation though so the earliest I could do anything is the week of 12 August.
