### 2019-08-02 #1262 @r-barnes
When I try `pip install -U Scyland3D` and run `Scyland3D` from the CLI, I get
```
Traceback (most recent call last):
  File "/home/rick/.local/bin/Scyland3D", line 7, in <module>
    from Scyland3D import main
ImportError: cannot import name 'main'
```
a similar error occurs when using `python3 setup.py install --user`
