### 2019-08-02 ybayle/Scyland3D#1 @r-barnes
The program doesn't run from the CLI. It fails with the warning:
```
Traceback (most recent call last):
  File "/home/rick/.local/bin/Scyland3D", line 7, in <module>
    from Scyland3D import main
ImportError: cannot import name 'main'
```

When run without arguments, my expectation is that the program will offer help:
```
python3 ./Scyland3D/Scyland3D.py
```
However, it instead processes a directory of example files. This is not helpful.

Consider [`argparse`](https://docs.python.org/3/library/argparse.html) as a more Pythonic way of parsing arguments, as suggested in the original review comment.
