### 2019-08-02 ybayle/Scyland3D#8 @r-barnes
This issue seems almost entirely unaddressed. Please comment your code.
