### 2019-08-02 ybayle/Scyland3D#8 @r-barnes
The docstrings are an excellent addition. However, certain parts of the code are still mysterious.

For instance, what does this equation

    t = float((D - A * x - B * y - C * z) / (A * A + B * B + C * C))

do?

A more liberal use of comments within the source itself will make it easier for you and others to modify the code later.
