### 2019-08-02 #1262 @r-barnes
I see the following issues:
 * [ ] Both `pip3 install -U Scyland3D` and `python3 setup.py install Scyland3D --user` lead to an unusable state: `ImportError: cannot import name 'main'`
 * [ ] Calling without command line arguments does not provide help. Instead, an example directory is parsed. This violates the principle of least surprise.
 * [ ] Argument parsing seems unnecessarily verbose. My first review suggested the `argparse` package. I'm not sure why `getopt` has been used.
 * [ ] Image in README appears without justification. It should be used to support the initial description of what the software does.
 * [ ] Internal comments are lacking near critical places. What, for instance, does `t = float((D - A * x - B * y - C * z) / (A * A + B * B + C * C))` do? Try to describe mathematical operations and other patches of non-contextualized code to orientate yourself and others. Particularly, [Issue 8](https://github.com/ybayle/Scyland3D/issues/8) seems disappointingly unaddressed.

Other than this, the situation is improved versus the original submission; however, my sense that the software solves too specific a problem in too specific a way remains.
