### 2019-08-06 #1262 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman I don't mind personally. But if there's anything more that would help you (like doing an official checklist), I'd be happy to do it. Should be trivial given I've already completed a pretty much full review and everything is fine from my end.
> I am on vacation though so the earliest I could do anything is the week of 12 August.

Thanks @patrikhuber I've just added you as an official reviewer. Since you did all the work and did effectively perform a review I think it is best to keep track of it officially like this. You should be able to tick the boxes at the top of this issue under your handle.
Let me know if you have any questions. Thanks again!
