### 2019-08-24 #1262 @Kevin-Mattheus-Moerman
@r-barnes I noticed the "Functionality" box is not ticked. Can you summarize where you think this needs improvement or state what is needed for you to tick this box? Thanks!
