### 2019-09-04 #1262 @Kevin-Mattheus-Moerman
@r-barnes can you summarize what points you feel should still be addressed? Thanks!

> @r-barnes I noticed the "Functionality" box is not ticked. Can you summarize where you think this needs improvement or state what is needed for you to tick this box? Thanks!
