### 2019-09-04 #1262 @patrikhuber
@Kevin-Mattheus-Moerman I've just done the checklist, apologies it took a while.
I haven't checked off two boxes as I am having some trouble running Scyland3D from pip - will open an issue in the repo.

Also one other thing I found, the README.md contains the sentence & link `More information can be found on the related article.` - The link is currently a 404, this needs to be updated, I guess once the paper has been accepted and published in JOSS.
