### 2019-09-04 ybayle/Scyland3D#10 @patrikhuber
Having trouble running Scyland3D when installed from pip
Hi!

If I clone the repo from git and go into the `Scyland3D\Scyland3D` directory on a command-line, I can successfully run the example and it creates the `.csv` file. However, when I create a new virtualenv and do `pip install Scyland3D`, then I can still `import` the module without error, but it doesn't seem to have any functions. `help(Scyland3D)` then looks like an empty module and trying to call any function fails, e.g. `AttributeError: module 'Scyland3D' has no attribute 'pts2csv'`.

I tried this on both Windows and on WSL Ubuntu 18.04.

Anything that I might be doing wrong here?
