### 2019-09-04 ybayle/Scyland3D#11 @patrikhuber
AssertionError when running tests
Hi,

When I clone the git repo and run the tests, I'm getting the following error:

```
(scyland3d-review) patrik:/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D$ ../test.sh
Testing call from python...
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
File successfully generated: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/landmarks.csv
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
File successfully generated: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/landmarks_reordered.csv
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
File successfully generated: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/landmarks_reversed.csv
Testing default behavior...
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/Scyland3D.py", line 486, in test_no_regression
    _validation_against_ref()
  File "/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/Scyland3D.py", line 391, in _validation_against_ref
    dirn + "landmarks.csv", dirn + "test/landmarks_ref.csv"
  File "/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/Scyland3D.py", line 383, in _same_file
    assert line_ref == line_test, "Invalid line " + line_ref
AssertionError: Invalid line example/200118G_canicula_mediterranean_40_female_mature_lower_tooth04_4.pts,23.760645,14.822544,26.607994,23.460169,14.72399,26.149445,23.619444,14.667845,26.251095,23.519606,14.463796,25.761469,23.799814,14.451308,26.218939,23.817883,14.327648,26.021515,24.05011,14.500256,26.583118,23.731268,14.851231,26.612024,23.693602,14.879261,26.590559,23.657339,14.899388,26.562233,23.620415,14.917179,26.531328,23.591503,14.916238,26.493502,23.580864,14.887245,26.456419,23.555262,14.877494,26.415607,23.52702,14.865196,26.368336,23.579859,14.780132,26.356867,23.546961,14.772367,26.300602,23.504704,14.757745,26.230452,23.531639,14.690864,26.178947,23.574799,14.58428,26.049805,23.669212,14.445662,26.010353,23.800825,14.386982,26.109592,23.858444,14.317473,26.077599,23.894411,14.332585,26.170313,23.92621,14.366543,26.286236,23.955145,14.364937,26.324955,23.982096,14.327062,26.290644,24.00683,14.300524,26.272785,24.027893,14.302449,26.299009,24.04743,14.317135,26.343693,24.0672,14.343269,26.407444,24.082382,14.387399,26.492682,24.071802,14.447412,26.555712,23.957222,14.536147,26.573864,23.874788,14.572289,26.542034,23.817228,14.613439,26.521942,23.779778,14.664021,26.510342,23.76063,14.742847,26.568264,200118G,canicula,mediterranean,40,female,mature,lower,tooth04,4

Testing call from the command line...
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
File successfully generated: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/landmarks.csv
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
Traceback (most recent call last):
  File "Scyland3D.py", line 490, in <module>
    main(sys.argv[1:])
  File "Scyland3D.py", line 374, in main
    verbose=verbose,
  File "Scyland3D.py", line 211, in pts2csv
    ), "Must supply order and order_factor."
AssertionError: Must supply order and order_factor.
Using default example directory: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/example/
File successfully generated: /mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/landmarks_reversed.csv
Testing default behavior...
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/Scyland3D.py", line 391, in _validation_against_ref
    dirn + "landmarks.csv", dirn + "test/landmarks_ref.csv"
  File "/mnt/c/Users/Patrik/Projects/Scyland3D-review/Scyland3D/Scyland3D/Scyland3D.py", line 383, in _same_file
    assert line_ref == line_test, "Invalid line " + line_ref
AssertionError: Invalid line example/200118G_canicula_mediterranean_40_female_mature_lower_tooth04_4.pts,23.760645,14.822544,26.607994,23.460169,14.72399,26.149445,23.619444,14.667845,26.251095,23.519606,14.463796,25.761469,23.799814,14.451308,26.218939,23.817883,14.327648,26.021515,24.05011,14.500256,26.583118,23.731268,14.851231,26.612024,23.693602,14.879261,26.590559,23.657339,14.899388,26.562233,23.620415,14.917179,26.531328,23.591503,14.916238,26.493502,23.580864,14.887245,26.456419,23.555262,14.877494,26.415607,23.52702,14.865196,26.368336,23.579859,14.780132,26.356867,23.546961,14.772367,26.300602,23.504704,14.757745,26.230452,23.531639,14.690864,26.178947,23.574799,14.58428,26.049805,23.669212,14.445662,26.010353,23.800825,14.386982,26.109592,23.858444,14.317473,26.077599,23.894411,14.332585,26.170313,23.92621,14.366543,26.286236,23.955145,14.364937,26.324955,23.982096,14.327062,26.290644,24.00683,14.300524,26.272785,24.027893,14.302449,26.299009,24.04743,14.317135,26.343693,24.0672,14.343269,26.407444,24.082382,14.387399,26.492682,24.071802,14.447412,26.555712,23.957222,14.536147,26.573864,23.874788,14.572289,26.542034,23.817228,14.613439,26.521942,23.779778,14.664021,26.510342,23.76063,14.742847,26.568264,200118G,canicula,mediterranean,40,female,mature,lower,tooth04,4
```

The `Invalid line ` error could in theory be a newline inconsistency since I'm using WSL - just a guess. The `AssertionError: Must supply order and order_factor.` error looks like something else.

Any idea what I might be doing wrong?
