### 2019-09-16 ybayle/Scyland3D#11 @Kevin-Mattheus-Moerman
@ybayle :wave: any updates on handling this issue?
