### 2019-10-17 282ec45
enhance assertion verbosity for tests
