### 2019-10-17 ybayle/Scyland3D#11 @ybayle
Thanks for raising this issue. I fixed `The AssertionError: Must supply order and order_factor.`. For the WSL related issue, I have added verbosity in the test to help figure out the problem. If you don't reproduce it with the last version, please let me know.
