### 2019-10-17 ybayle/Scyland3D#10 @ybayle
Thank for raising this issue. I have updated the Readme to indicate that it is necessary to use `from Scyland3D import Scyland3D` in order to avoid this kind of problem. Could you confirm that this solve the issue?
