### 2019-10-17 2de44ca
display usage when calling without arg
