### 2019-10-19 01914c0
Add image descrition in the Readme
