### 2019-10-19 ebc20cb
add comment about the symmetry point
