### 2019-10-19 c75e13c
add comments for clarity
