### 2019-10-19 #1262 @ybayle
@patrikhuber Thanks for your comments. I solved ybayle/Scyland3D#10 and ybayle/Scyland3D#11. Tell me if everything is ok for you in the new version 1.0.17.

```
Also one other thing I found, the README.md contains the sentence & link More information can be found on the related article. - The link is currently a 404, this needs to be updated, I guess once the paper has been accepted and published in JOSS.
```

That's it, it will be updated upon acceptance.

@r-barnes Thanks for your suggestions that have all been added. N.B.: the use of argparse is far better than optget!

@Kevin-Mattheus-Moerman Everything is ok and up to date on my side, except for the new version. Could you update it on your side please? Thanks!
