### 2019-10-19 ybayle/Scyland3D#10 @patrikhuber
I can confirm that using `from Scyland3D import Scyland3D` fixes this issue.

I don't know too much about packaging Python modules, but this seems a bit unusual to me. Though I've seen this syntax before of course, I would expect that a simple `import Scyland3D` should work too? Or subsequently, `import Scyland3D as S` should work too (but it doesn't). For example it's common practice to do `import tensorflow as tf` so I would expect that `import Scyland3D as S` needs to work as well?
