### 2019-10-19 ybayle/Scyland3D#10 @patrikhuber
Your own tests also still fails since it uses `import Scyland3D`: https://github.com/ybayle/Scyland3D/blob/c75e13cb0f372a9a5a271931fbc0624d942435b9/test.sh#L2

Instead of just "fixing" this to `from Scyland3D import Scyland3D` it would be really good to know why the usual method doesn't work when it does work for lots of other libraries (and when Scyland3D is used from GitHub).
