### 2019-10-19 ybayle/Scyland3D#11 @patrikhuber
The `Must supply order and order_factor.` error seems gone.

There's a different error now:

```
(scyland3d) C:\Users\Patrik>python
Python 3.7.3 (v3.7.3:ef4ec6ed12, Mar 25 2019, 22:22:05) [MSC v.1916 64 bit (AMD64)] on win32
>>> from Scyland3D import Scyland3D
>>> Scyland3D.test_no_regression()
File successfully generated: C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\landmarks.csv
File successfully generated: C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\landmarks_reordered.csv
File successfully generated: C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\landmarks_reversed.csv
Testing default behavior...
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\Scyland3D.py", line 459, in test_no_regression
    _validation_against_ref()
  File "C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\Scyland3D.py", line 351, in _validation_against_ref
    dirn + "landmarks.csv", dirn + "test/landmarks_ref.csv"
  File "C:\Users\Patrik\Envs\scyland3d\lib\site-packages\Scyland3D\Scyland3D.py", line 334, in _same_file
    assert line_ref == line_test, "Invalid line generated:\n" + line_test + "\nthat is different from the reference file:\n" + line_ref
AssertionError: Invalid line generated:
example/200118G_canicula_mediterranean_40_female_mature_lower_tooth09_4.pts,25.628664,11.627904,23.958033,25.294479,11.766243,23.443439,25.416937,11.636269,23.504972,25.285938,11.601955,23.1847,25.480515,11.458488,23.40572,25.424305,11.401969,23.213655,25.720415,11.227379,23.608505,25.605238,11.68239,23.968342,25.581594,11.735225,23.951746,25.55871,11.775233,23.927303,25.535694,11.803462,23.898464,25.509012,11.828192,23.866955,25.481352,11.8446,23.82785,25.486269,11.775678,23.780958,25.443384,11.805834,23.732676,25.394228,11.839259,23.66806,25.410519,11.719546,23.634127,25.351175,11.754179,23.546564,25.337692,11.711373,23.459862,25.341736,11.625697,23.339272,25.361479,11.536088,23.272907,25.453016,11.429392,23.314713,25.506357,11.353887,23.309586,25.548668,11.321219,23.384398,25.550634,11.305209,23.380384,25.552917,11.292115,23.362726,25.55512,11.280545,23.327475,25.565844,11.266936,23.305206,25.586967,11.251483,23.31579,25.614349,11.235746,23.352438,25.64748,11.214075,23.401062,25.69471,11.177201,23.487968,25.711798,11.199377,23.571569,25.649658,11.312396,23.602541,25.589733,11.408429,23.600561,25.570297,11.46718,23.659069,25.563253,11.537304,23.721621,25.585634,11.584576,23.828579,200118G,canicula,mediterranean,40,female,mature,lower,tooth09,4

that is different from the reference file:

>>>
```

This is on regular Windows (haven't tested WSL again so far). I still suspect it may have to do with line endings.
