### 2019-10-19 #1262 @patrikhuber
@ybayle @Kevin-Mattheus-Moerman I've retested today and updated the issues.
