### 2019-10-19 ybayle/Scyland3D#11 @patrikhuber
Got an error in WSL as well:

```
(scyland3d) patrik:~/scyland3d-review/Scyland3D/Scyland3D$ ../test.sh
Testing call from python...
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks.csv
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks_reordered.csv
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks_reversed.csv
Testing default behavior...
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/patrik/scyland3d-review/Scyland3D/Scyland3D/Scyland3D.py", line 459, in test_no_regression
    _validation_against_ref()
  File "/home/patrik/scyland3d-review/Scyland3D/Scyland3D/Scyland3D.py", line 351, in _validation_against_ref
    dirn + "landmarks.csv", dirn + "test/landmarks_ref.csv"
  File "/home/patrik/scyland3d-review/Scyland3D/Scyland3D/Scyland3D.py", line 334, in _same_file
    assert line_ref == line_test, "Invalid line generated:\n" + line_test + "\nthat is different from the reference file:\n" + line_ref
AssertionError: Invalid line generated:
example/200118G_canicula_mediterranean_40_female_mature_lower_tooth09_4.pts,25.628664,11.627904,23.958033,25.294479,11.766243,23.443439,25.416937,11.636269,23.504972,25.285938,11.601955,23.1847,25.480515,11.458488,23.40572,25.424305,11.401969,23.213655,25.720415,11.227379,23.608505,25.605238,11.68239,23.968342,25.581594,11.735225,23.951746,25.55871,11.775233,23.927303,25.535694,11.803462,23.898464,25.509012,11.828192,23.866955,25.481352,11.8446,23.82785,25.486269,11.775678,23.780958,25.443384,11.805834,23.732676,25.394228,11.839259,23.66806,25.410519,11.719546,23.634127,25.351175,11.754179,23.546564,25.337692,11.711373,23.459862,25.341736,11.625697,23.339272,25.361479,11.536088,23.272907,25.453016,11.429392,23.314713,25.506357,11.353887,23.309586,25.548668,11.321219,23.384398,25.550634,11.305209,23.380384,25.552917,11.292115,23.362726,25.55512,11.280545,23.327475,25.565844,11.266936,23.305206,25.586967,11.251483,23.31579,25.614349,11.235746,23.352438,25.64748,11.214075,23.401062,25.69471,11.177201,23.487968,25.711798,11.199377,23.571569,25.649658,11.312396,23.602541,25.589733,11.408429,23.600561,25.570297,11.46718,23.659069,25.563253,11.537304,23.721621,25.585634,11.584576,23.828579,200118G,canicula,mediterranean,40,female,mature,lower,tooth09,4

that is different from the reference file:
example/200118G_canicula_mediterranean_40_female_mature_lower_tooth04_4.pts,23.760645,14.822544,26.607994,23.460169,14.72399,26.149445,23.619444,14.667845,26.251095,23.519606,14.463796,25.761469,23.799814,14.451308,26.218939,23.817883,14.327648,26.021515,24.05011,14.500256,26.583118,23.731268,14.851231,26.612024,23.693602,14.879261,26.590559,23.657339,14.899388,26.562233,23.620415,14.917179,26.531328,23.591503,14.916238,26.493502,23.580864,14.887245,26.456419,23.555262,14.877494,26.415607,23.52702,14.865196,26.368336,23.579859,14.780132,26.356867,23.546961,14.772367,26.300602,23.504704,14.757745,26.230452,23.531639,14.690864,26.178947,23.574799,14.58428,26.049805,23.669212,14.445662,26.010353,23.800825,14.386982,26.109592,23.858444,14.317473,26.077599,23.894411,14.332585,26.170313,23.92621,14.366543,26.286236,23.955145,14.364937,26.324955,23.982096,14.327062,26.290644,24.00683,14.300524,26.272785,24.027893,14.302449,26.299009,24.04743,14.317135,26.343693,24.0672,14.343269,26.407444,24.082382,14.387399,26.492682,24.071802,14.447412,26.555712,23.957222,14.536147,26.573864,23.874788,14.572289,26.542034,23.817228,14.613439,26.521942,23.779778,14.664021,26.510342,23.76063,14.742847,26.568264,200118G,canicula,mediterranean,40,female,mature,lower,tooth04,4

Testing call from the command line...
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks.csv
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks_reordered.csv
File successfully generated: /home/patrik/scyland3d-review/Scyland3D/Scyland3D/landmarks_reversed.csv
Testing default behavior...
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/patrik/scyland3d-review/Scyland3D/Scyland3D/Scyland3D.py", line 351, in _validation_against_ref
    dirn + "landmarks.csv", dirn + "test/landmarks_ref.csv"
  File "/home/patrik/scyland3d-review/Scyland3D/Scyland3D/Scyland3D.py", line 334, in _same_file
    assert line_ref == line_test, "Invalid line generated:\n" + line_test + "\nthat is different from the reference file:\n" + line_ref
AssertionError: Invalid line generated:
example/200118G_canicula_mediterranean_40_female_mature_lower_tooth09_4.pts,25.628664,11.627904,23.958033,25.294479,11.766243,23.443439,25.416937,11.636269,23.504972,25.285938,11.601955,23.1847,25.480515,11.458488,23.40572,25.424305,11.401969,23.213655,25.720415,11.227379,23.608505,25.605238,11.68239,23.968342,25.581594,11.735225,23.951746,25.55871,11.775233,23.927303,25.535694,11.803462,23.898464,25.509012,11.828192,23.866955,25.481352,11.8446,23.82785,25.486269,11.775678,23.780958,25.443384,11.805834,23.732676,25.394228,11.839259,23.66806,25.410519,11.719546,23.634127,25.351175,11.754179,23.546564,25.337692,11.711373,23.459862,25.341736,11.625697,23.339272,25.361479,11.536088,23.272907,25.453016,11.429392,23.314713,25.506357,11.353887,23.309586,25.548668,11.321219,23.384398,25.550634,11.305209,23.380384,25.552917,11.292115,23.362726,25.55512,11.280545,23.327475,25.565844,11.266936,23.305206,25.586967,11.251483,23.31579,25.614349,11.235746,23.352438,25.64748,11.214075,23.401062,25.69471,11.177201,23.487968,25.711798,11.199377,23.571569,25.649658,11.312396,23.602541,25.589733,11.408429,23.600561,25.570297,11.46718,23.659069,25.563253,11.537304,23.721621,25.585634,11.584576,23.828579,200118G,canicula,mediterranean,40,female,mature,lower,tooth09,4

that is different from the reference file:
example/200118G_canicula_mediterranean_40_female_mature_lower_tooth04_4.pts,23.760645,14.822544,26.607994,23.460169,14.72399,26.149445,23.619444,14.667845,26.251095,23.519606,14.463796,25.761469,23.799814,14.451308,26.218939,23.817883,14.327648,26.021515,24.05011,14.500256,26.583118,23.731268,14.851231,26.612024,23.693602,14.879261,26.590559,23.657339,14.899388,26.562233,23.620415,14.917179,26.531328,23.591503,14.916238,26.493502,23.580864,14.887245,26.456419,23.555262,14.877494,26.415607,23.52702,14.865196,26.368336,23.579859,14.780132,26.356867,23.546961,14.772367,26.300602,23.504704,14.757745,26.230452,23.531639,14.690864,26.178947,23.574799,14.58428,26.049805,23.669212,14.445662,26.010353,23.800825,14.386982,26.109592,23.858444,14.317473,26.077599,23.894411,14.332585,26.170313,23.92621,14.366543,26.286236,23.955145,14.364937,26.324955,23.982096,14.327062,26.290644,24.00683,14.300524,26.272785,24.027893,14.302449,26.299009,24.04743,14.317135,26.343693,24.0672,14.343269,26.407444,24.082382,14.387399,26.492682,24.071802,14.447412,26.555712,23.957222,14.536147,26.573864,23.874788,14.572289,26.542034,23.817228,14.613439,26.521942,23.779778,14.664021,26.510342,23.76063,14.742847,26.568264,200118G,canicula,mediterranean,40,female,mature,lower,tooth04,4

(scyland3d) patrik:~/scyland3d-review/Scyland3D/Scyland3D$
```
