### 2019-10-20 #1262 @ybayle
> have these issues raised by @r-barnes been addressed

@Kevin-Mattheus-Moerman Yes but both reviewers raised new issues that I'm tackling.
