### 2019-10-20 ybayle/Scyland3D#10 @ybayle
> I can confirm that using from Scyland3D import Scyland3D fixes this issue.

:+1:

> I don't know too much about packaging Python modules, but this seems a bit unusual to me. Though I've seen this syntax before of course, I would expect that a simple import Scyland3D should work too? Or subsequently, import Scyland3D as S should work too (but it doesn't). For example it's common practice to do import tensorflow as tf so I would expect that import Scyland3D as S needs to work as well?

The keyword `as` serves a different purpose: it is just an alias. It actually works fine if you write:
```
from Scyland3D import Scyland3D as S
S.pts2csv()
```

Moreover here, `Scyland3D` is the name of the package and the name of the module. That's why you need to use `from Scyland3D import Scyland3D`, which is common practice for this kind of small packages. For tensorflow there are multiple layers of package/module/functionalities so that's why you encounter import such as `from tensorflow.python.keras.layers import Input, Dense`.

> Your own tests also still fails since it uses import Scyland3D:

On my machine, the test works because the test.sh script considers the Scyland3D module directly as it is stored in the same folder.
If the test fails on your machine could you describe more precisely the error you encounter?
