### 2019-10-20 a01a14f
fix #8 by adding more comments describing the code
