### 2019-10-20 ybayle/Scyland3D#8 @ybayle
I pushed a new commit that enhance the degree of verbosity of the comments in the code in answer to this issue and to the comments detailed here: https://github.com/ybayle/Scyland3D/commit/c75e13cb0f372a9a5a271931fbc0624d942435b9#comments
Could you confirm that this issue is solved?
