### 2019-10-20 ybayle/Scyland3D#11 @ybayle
I fixed the issue on Windows here, could you confirm everything is working fine on your side?
