### 2019-10-20 #1262 @ybayle
@Kevin-Mattheus-Moerman I have addressed all reviewers comments, suggestions and fixed every bug raised by the reviewers. I have added a new release on the main repo, could you update again the version to v1.0.18 please? Thanks! I'm waiting for confirmation from @r-barnes and @patrikhuber that everything is ok for the v1.0.18 release.
