### 2019-10-21 ybayle/Scyland3D#8 @patrikhuber
a01a14f2657134fd2ae7a91cc9fa9bf3e04a906b improved the readability / understanding of the code a lot! 👍

It would probably be good to hear @r-barnes's comment about it but I have no further comments.
