### 2019-10-29 ybayle/Scyland3D#11 @patrikhuber
More tests seem to pass now but I still got an error:
(this is in a  WSL shell so it might again have to do with line endings - or not, as two floating point values seem to be quite different)

```
$ python
Python 3.6.8 (default, Oct  7 2019, 12:59:55)
[GCC 8.3.0] on linux
>>> from Scyland3D import Scyland3D
>>> Scyland3D.test_no_regression()
File successfully generated: /home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/landmarks.csv
File successfully generated: /home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/landmarks_reordered.csv
File successfully generated: /home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/landmarks_reversed.csv
Testing default behavior...
Ok!
Testing reordering...
Ok!
Testing reversing...
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/Scyland3D.py", line 473, in test_no_regression
    _validation_against_ref()
  File "/home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/Scyland3D.py", line 375, in _validation_against_ref
    dirn + "landmarks_reversed.csv", dirn + "test/landmarks_reversed_ref.csv"
  File "/home/patrik/.virtualenvs/scyland3d/lib/python3.6/site-packages/Scyland3D/Scyland3D.py", line 348, in _same_file
    assert line_ref == line_test, "Invalid line generated:\n" + line_test + "\nthat is different from the reference file:\n" + line_ref
AssertionError: Invalid line generated:
example/200118H_canicula_mediterranean_47_female_mature_upper_tooth04_3.pts,15.250192222098738,21.508294685741003,28.678250805861243,15.538475872969613,21.315691297389527,28.44522452981047,15.412222002840693,21.26600127842893,28.544590769400006,15.716011901860037,20.940988874361956,28.232551859944113,15.321003219633203,20.929705218359096,28.671938345414056,15.385567752898783,20.773448718255768,28.62005217882164,15.044217842301663,20.862119786950778,29.01015235346634,15.280107387672018,21.526176742024028,28.66775158604194,15.32016829776629,21.531716073352488,28.640497426508908,15.356263793839194,21.51848580106889,28.612198967265464,15.38706693132819,21.506442769862563,28.589136153946235,15.41582177874779,21.49417770434044,28.56699502742536,15.43422670225145,21.471456221987697,28.553802789271113,15.44618470828529,21.44582790694996,28.546635853173925,15.461553248298133,21.423917007187256,28.534659814494024,15.478662622688843,21.40206336691809,28.519027844466407,15.501531922335836,21.379159154157453,28.493967367426425,15.531206075844668,21.349305640929018,28.455807673160695,15.485155720083032,21.29784494290733,28.48724824579142,15.556924362192811,21.117649540829568,28.42137611223139,15.50403626716043,20.94949452454902,28.48056039616332,15.3549235878322,20.853872688018733,28.644836365976715,15.354441539203991,20.7712248168309,28.660442274705144,15.318230969761704,20.7869293325365,28.708930605756212,15.285078474811566,20.803707476108308,28.754579286546186,15.254424366245278,20.82061625522485,28.79583384298295,15.232625333366101,20.809478775537595,28.81873616054578,15.216799746854711,20.764003550805196,28.827747724899883,15.189599739565143,20.764114082361466,28.857830582796407,15.159992006074468,20.781359000017076,28.891251832901638,15.130997634685242,20.793259298842436,28.924030919872003,15.100843105169742,20.80799305027017,28.9609132676764,15.07096531815483,20.833419079058142,28.99349293847223,15.112607287397035,20.949018519924085,28.965924240278213,15.186083400519212,21.033059514371658,28.863261074630934,15.242057300694782,21.11957457419216,28.78464398025325,15.27852206033822,21.216743637876064,28.72257883244919,15.27852649703689,21.35115009152016,28.688980969881268,200118H,canicula,mediterranean,47,female,mature,upper,tooth04,3

that is different from the reference file:
example/200118H_canicula_mediterranean_47_female_mature_upper_tooth04_3.pts,15.250192222093075,21.508294685740747,28.678250805856255,15.538475872963792,21.315691297389172,28.44522452980537,15.412222002834772,21.26600127842862,28.5445907693948,15.716011901853829,20.940988874361615,28.23255185993866,15.321003219626819,20.92970521835876,28.671938345408446,15.385567752892229,20.773448718255406,28.620052178815882,15.04421784229508,20.86211978695041,29.01015235346056,15.28010738766638,21.526176742023747,28.66775158603698,15.32016829776067,21.53171607335219,28.64049742650397,15.356263793833568,21.518485801068575,28.61219896726052,15.387066931322554,21.506442769862232,28.589136153941293,15.415821778742147,21.494177704340103,28.566995027420415,15.434226702245786,21.47145622198735,28.553802789266147,15.446184708279594,21.445827906949607,28.546635853168933,15.461553248292415,21.423917007186894,28.534659814489018,15.478662622683105,21.402063366917726,28.519027844461384,15.501531922330077,21.379159154157087,28.49396736742138,15.531206075838886,21.34930564092866,28.455807673155633,15.48515572007717,21.297844942906995,28.487248245786276,15.556924362186754,21.117649540829213,28.421376112226074,15.504036267154138,20.94949452454867,28.480560396157795,15.354923587825736,20.853872688018384,28.644836365971038,15.354441539197417,20.771224816830536,28.660442274699374,15.318230969755131,20.786929332536122,28.708930605750446,15.285078474804996,20.803707476107924,28.754579286540423,15.254424366238714,20.820616255224458,28.795833842977196,15.232625333359518,20.809478775537205,28.818736160540006,15.216799746848071,20.764003550804816,28.827747724894056,15.189599739558494,20.764114082361086,28.857830582790573,15.15999200606783,20.781359000016696,28.89125183289581,15.1309976346786,20.793259298842056,28.924030919866173,15.10084310516311,20.807993050269786,28.96091326767058,15.070965318148213,20.83341907905776,28.993492938466424,15.112607287390576,20.94901851992368,28.965924240272553,15.186083400512892,21.033059514371274,28.863261074625395,15.242057300688595,21.119574574191795,28.78464398024782,15.278522060332175,21.216743637875723,28.722578832443883,15.278526497031024,21.35115009151986,28.688980969876113,200118H,canicula,mediterranean,47,female,mature,upper,tooth04,3

>>>
```
