### 2019-10-29 ybayle/Scyland3D#10 @patrikhuber
> Moreover here, Scyland3D is the name of the package and the name of the module. That's why you need to use from Scyland3D import Scyland3D, which is common practice for this kind of small packages.

I see! As I said I am not too familiar with Python packaging, so I wasn't aware of the package/module difference. But that makes sense.

>> If the test fails on your machine could you describe more precisely the error you encounter?

I think it only failed because I tried to run the test manually, using the Scyland3D module from `pip`, and not using the one from the directory (where the test script is located). So I think it's all good.

I think it's all fine here, so I'm going to close the issue.
