### 2019-10-29 ybayle/Scyland3D#11 @ybayle
All values seem to be different at 10^-11 this might be platform specific and due to rounding errors. Depending on the python version and OS, the minimum epsilon that can be represented is greater than 1e-15 obtained with:
```
import numpy as np
print(np.finfo(np.float))
```
