### 2019-10-29 ybayle/Scyland3D#11 @patrikhuber
Would it make sense to check for an eps of 10^-9 or 10^-10 or so in the tests? As far as I understand, reversing involves some mathematical operations (right?), so a difference this small is expected?
