### 2019-10-29 ybayle/Scyland3D#11 @ybayle
This kind of difference is indeed expected. I'll fix this.
