### 2019-10-29 ybayle/Scyland3D#11 @ybayle
Could you check after a `git pull` that the last commits solve your issue? Thanks!
If so, I'll update the software version on pip and resubmit it to the JOSS editor.
