### 2019-10-29 ybayle/Scyland3D#13 @patrikhuber
Mistake / old information in Readme for command-line arguments
I think I found another piece of information in the Readme that doesn't match the actual command-line parameters (perhaps you forgot to update those when you switched to the new cmdline options module).

Readme says:

> Command line: python -m Scyland3D.Scyland3D -v False for removing the display of information to the console.

Output:

```
python -m Scyland3D.Scyland3D -v False
usage: Scyland3D.py [-h] [-i INDIR] [-o OUTDIR] [-m MIRROR_FACTOR] [-r ORDER]
                    [-f ORDER_FACTOR] [-n FEATURE_NAMES] [-v]
Scyland3D.py: error: unrecognized arguments: False
```

Could you also check all the other examples in the readme please whether they work as expected?
