### 2019-10-29 ybayle/Scyland3D#11 @patrikhuber
I can confirm that all tests pass now!
