### 2019-10-30 a78e160
update all examples proposed in the readme
