### 2019-10-30 ybayle/Scyland3D#13 @ybayle
I updated all examples from command line and from python. Could you tell me if everything is ok on your side?
