### 2019-10-30 ybayle/Scyland3D#13 @patrikhuber
A couple more questions:

> This is only valid if the keyword upper is present in the file name.

I'm a bit confused what that exactly means. In the filename? Or within one of the .pts files in the directory given with `-i`? In case of the latter, I am still confused because none of these files contains the string "upper" (or I am not finding it / looking in the wrong place?)

Also when I run the given command, `python -m Scyland3D.Scyland3D -i "path/" -f "upper" -r "36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 37"`, on the `example/` directory, a file `landmarks_reordered.csv` gets written.
If I replace `-f "upper"` by `-f "lower"`, it writes a file `landmarks.csv`. Why this unexpected and (for me) unintuitive difference?

The other commands exhibit the same behaviour, e.g. with `landmarks_reversed.csv` etc.

> Command line: python -m Scyland3D.Scyland3D -i "path/" -m "upper" will mirror in 3D the files containing the keyword upper before processing them with the remaining lower items as depicted in the figure below:

The figure is no longer below that text. Also if you give a non-existent argument like "asdf", there will be no error. Is that expected?
The text description is also not 100% clear to me. "will mirror in 3D the files containing the keyword upper before processing them with the remaining lower items"? "...before processing them with the remaining..."?
