### 2019-10-30 ybayle/Scyland3D#13 @ybayle
`AssertionError: The order_factor (asdf) provided has not been found in any file names.`
