### 2019-10-31 ybayle/Scyland3D#13 @patrikhuber
> Command line: python -m Scyland3D.Scyland3D -i "path/" -m "upper" will mirror in 3D the files containing the keyword upper before processing them with the lower items as depicted in the figure above.

I am still not sure the second part (after the "before...") is clear. If I understand correctly, what you want to say is "...before processing the remaining files (without mirroring them).". Correct? And the figure depicts the mirroring, so it only applies to the first part of the sentence?

---

If you specify part of a string, e.g. `... -f "u"`, then it runs as well. I guess because it matches the "u" to "upper"? Is this expected, or would it not be better to search for an _exact_ match? Otherwise if a file contains `_lower_` in the filename but contains an `u` somewhere else (maybe you're procesing mo**u**ths and not teeth!), it would match it and reorder as well?

Also I understand now that you're really referring to the filename, and not to the file content. Encoding information in the filename is a tricky business, in my opinion. It might be unavoidable though if you don't have control over the content of the .pts file / format. In that case, you should just document it better that the **filename** of the files actually matter and that they have to have a specific format (e.g. separated by `_` or `-` or something so that exact matches can be produced).

An even better solution might be that you take as input to the script not a whole folder, but only a single file (usually this is better practice anyway), and then the order arguments can be given directly, without having to parse filenames. If someone then needs to process a whole folder, they would write their own script around yours that handles the folder processing and filename convention. (You could also supply this, separately).

Anyway this decision I would leave up to you, just some input to consider from my side.
