### 2019-11-03 c57852b
enhance readme for #13
