### 2019-11-03 ybayle/Scyland3D#13 @patrikhuber
> Command line: `python -m Scyland3D.Scyland3D -i "path/" -m "upper"` will mirror in 3D -as depicted in the figure above- the landmarks stored in `.pts` file with the keyword `upper` in their filename. Then, it will export to a `.csv` file those processed landmarks and also the unprocessed landmarks from `.pts` file that do not contain the keyword `upper` in their filename.

I think this sentence is more clear but a bit convoluted now. I am not a native speaker either but how about this slight modification?

> Command line: `python -m Scyland3D.Scyland3D -i "path/" -m "upper"` will mirror in 3D the landmarks stored in `.pts` files with the keyword `upper` in their filename. Then, it will export to a `.csv` file those mirrored landmarks and also the unprocessed landmarks from `.pts` files that do not contain the keyword `upper` in their filename. The figure above depicts the mirroring.

---

> I consider this less dangerous than having a typo in a command line like rm filename.txt.

Well certainly deleting files by accident can be dangerous, especially if one doesn't have a backup ;-)
But I would argue that these sort of "silent mistakes" like here, are potentially more dangerous, particularly in research environments. You will run a command, often on several files, you will not hand-check all the resulting files usually, and you may never discover that some points that you thought got mirrored actually didn't, or the other way round.

Personally, I think the Scyland3D documentation is still not clear enough to allow someone *not* already familiar with it to understand how all this filename-stuff and the order-factor etc. works, and what the potential gotchas are. If we didn't have this discussion, I don't think I would be able to confidently use it, just from reading the Readme.

Anyway I will leave it at that - I feel like some of that might just be my personal opinion/preference, and I feel like my job as reviewer is done. It might be perfectly clear to other people and I know that A) reviewers aren't always right and B) authors can't (or don't want to) always fully please all reviewers :-).

Thanks for all the explanations and I'm sure they'll be helpful to your other users!
