### 2019-11-03 #1262 @patrikhuber
Just also mentioning https://github.com/ybayle/Scyland3D/issues/13 here as it is/was part of this review.

I would now consider all my issues closed and my review done.

Note that one thing I haven't done is to verify that the output of the processing is actually correct, i.e. that the reversed landmarks are actually correctly reversed or the re-odered ones correctly re-ordered. This would require more work and writing some plotting/visualisation scripts. There are unit tests in Scyland3D for these functions though and these pass on my machine.

I also still have my original comment that the software is extremely specific, both in what it does as well as the domain that it applies to. But that's not something that can be changed, it is what it is. I believe it is the Editor's decision whether this fits into JOSS.

Let me know if anything else is needed from my side.

All the best!
