### 2019-11-03 #1262 @ybayle
Thank you @patrikhuber for all your helpful comments about the code and the documentation that helped improve a lot our submission.
@Kevin-Mattheus-Moerman: I'm waiting for the validation of @r-barnes for my fixes on the last two issues he raised.
