### 2019-11-03 #1262 @ybayle
@Kevin-Mattheus-Moerman I have pushed v1.0.19 on pypi that fix the issues from @patrikhuber. Could you update the version accordingly?
