### 2019-12-26 #1262 @Kevin-Mattheus-Moerman
@r-barnes when you are ready can you check the fixes implemented. Let me know if you need more time. Let us know if anything else is stopping you from ticking that last box. Thanks! :christmas_tree:
