### 2020-01-20 #1262 @patrikhuber
@Kevin-Mattheus-Moerman I believe my comments in this thread sum it up pretty nicely, particularly my last one [[link](https://github.com/openjournals/joss-reviews/issues/1262#issuecomment-549161575)] should answer your question.
