### 2020-01-27 ybayle/Scyland3D#8 @r-barnes
It is much better.

Random aside. [This line](https://github.com/ybayle/Scyland3D/blob/a01a14f2657134fd2ae7a91cc9fa9bf3e04a906b/Scyland3D/Scyland3D.py#L304) _might_ be simplified through the use of regular expressions, but I'm not sure.
