### 2020-01-27 #1262 @r-barnes
@Kevin-Mattheus-Moerman : Things look largely good to me. The software/repo seem to be in good shape. The readability and documentation of the code is sufficient. The pip installation works. I remain concerned about the specificity of the software, but leave that to the editors.
