### 2020-01-28 #1262 @Kevin-Mattheus-Moerman
Thanks @r-barnes @patrikhuber for your additional comments and all the help with this review! I understand the concerns about the specificity of the software. Our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) do not currently list restrictions in relation to how extremely specific/niche software might be. We've had to reject submission in the past for being too specific/niche but only because they were so exotic or specialized that we were unable to find any suitable reviewers. In this case this what not the problem. I suppose perhaps if software is too specific it might not be classified as a "significant contribution" (which is in the guidelines). However in this case I feel this software offers a significant enough contribution to warrant publication. We may reflect on our submission requirements/guidelines though to perhaps include statements on how "specific/niche" software is.
Thanks again for your valuable contributions here!
