### 2020-01-29 #1262 @Kevin-Mattheus-Moerman
@ybayle I will now proceed to process this submission for acceptance in JOSS.

At this point can you:
- [ ] Archive the software on ZENODO and report back the DOI of the archived version here? Please ensure the title and author list of the ZENODO archived version match those of the paper (you may need to edit this information manually) .
- [ ] Can you inform me of the version tag that matches that archived version? Is it still at `v1.0.19`?
- [ ] This work is about to be accepted in JOSS. Can you please proofread the paper yourself one last time?

About the paper:
- [ ] the paper contains both `csv` (e.g. in `... to a csv format`) and `CSV`. Can you check you use `CSV` consistently throughout the text.
- [ ] In use `noise` not `noises` in `...may bring noises in the results.`
- [ ] Change `Presence of duplicates landmark...` to `Presence of duplicate landmarks...`
- [ ] Rephrase `...the two reviewers that helped improved a lot the code and the documentation.` to `...the two reviewers who helped improve the code and the documentation a lot.`
