### 2020-02-04 #1262 @ybayle
We proofread the paper again and generated a new version with the latest fixes: v1.1.0 cf https://pypi.org/project/Scyland3D/1.1.0/.
The Zenodo DOI is https://doi.org/10.5281/zenodo.3637546 but is not accessible yet. It's the first time that I used Zenodo, do I have to wait few days before Zenodo displays correctly every information?
