### 2020-02-05 #1262 @ybayle
@Kevin-Mattheus-Moerman
The generated article proof and the Zenodo are ok. What is the next step?
